+++ src/com/maccasoft/propeller/expressions/ContextLiteral.java
@@ -15,7 +15,7 @@ public class ContextLiteral extends Literal {
 
     @Override
     public String toString() {
-        return Integer.toString(getNumber().intValue());
+        return Long.toString(getNumber().longValue());
     }
 
 }
